+++ debugger/Dockerfile.dev
@@ -38,7 +38,7 @@ RUN . "$NVM_DIR/nvm.sh" && nvm alias default v${NODE_VERSION}
 ENV PATH="/root/.nvm/versions/node/v${NODE_VERSION}/bin/:${PATH}"
 RUN node --version
 RUN npm --version
-# RUN npm install -g nodemon
+RUN npm install -g nodemon
 
 WORKDIR /app
