--- debugger/samples/.gitignore
@@ -1,12 +1,0 @@
-linkedlist/main1
-linkedlist/main1.o
-linkedlist/main2
-linkedlist/main2.o
-linkedlist/main3
-linkedlist/main3.o
-linkedlist/linkedlist
-linkedlist/linkedlist.o
-program
-program.o
-buggy
-buggy.o
